+++ src/frontend/js/components/CourseGlimpseList/index.tsx
@@ -2,8 +2,7 @@ import { defineMessages, FormattedMessage } from 'react-intl';
 
 import { APIResponseListMeta } from 'types/api';
 import { CommonDataProps } from 'types/commonDataProps';
-import { Course } from 'types/Course';
-import { CourseGlimpse } from '../CourseGlimpse';
+import { CourseGlimpse, CourseGlimpseCourse } from 'components/CourseGlimpse';
 
 const messages = defineMessages({
   courseCount: {
@@ -22,43 +21,58 @@ const messages = defineMessages({
 });
 
 interface CourseGlimpseListProps {
-  courses: Course[];
-  meta: APIResponseListMeta;
+  courses: CourseGlimpseCourse[];
+  meta?: APIResponseListMeta;
+  className?: string;
 }
 
 export const CourseGlimpseList = ({
   context,
   courses,
   meta,
+  className,
 }: CourseGlimpseListProps & CommonDataProps) => {
+  const containerClassnames = ['course-glimpse-list'];
+  if (className) {
+    containerClassnames.push(className);
+  }
+
   return (
-    <div className="course-glimpse-list">
-      <div
-        className="offscreen"
-        data-testid="course-glimpse-sr-count"
-        aria-live="polite"
-        aria-atomic="true"
-      >
-        <FormattedMessage
-          {...messages.offscreenCourseCount}
-          values={{
-            courseCount: meta.total_count,
-          }}
-        />
-      </div>
-      <div className="course-glimpse-list__count" aria-hidden="true">
-        <FormattedMessage
-          {...messages.courseCount}
-          values={{
-            courseCount: meta.total_count,
-            end: meta.offset + meta.count,
-            start: meta.offset + 1,
-          }}
-        />
-      </div>
-      {courses.map(
-        (course) => course && <CourseGlimpse context={context} course={course} key={course.code} />,
+    <div className={containerClassnames.join(' ')}>
+      {meta && (
+        <div className="course-glimpse-list__header">
+          <div
+            className="offscreen"
+            data-testid="course-glimpse-sr-count"
+            aria-live="polite"
+            aria-atomic="true"
+          >
+            <FormattedMessage
+              {...messages.offscreenCourseCount}
+              values={{
+                courseCount: meta.total_count,
+              }}
+            />
+          </div>
+          <div className="course-glimpse-list__count" aria-hidden="true">
+            <FormattedMessage
+              {...messages.courseCount}
+              values={{
+                courseCount: meta.total_count,
+                end: meta.offset + meta.count,
+                start: meta.offset + 1,
+              }}
+            />
+          </div>
+        </div>
       )}
+      <div className="course-glimpse-list__content">
+        {courses.map((course) => (
+          <CourseGlimpse context={context} course={course} key={course.code} />
+        ))}
+      </div>
     </div>
   );
 };
+
+export { getCourseGlimpsListProps } from './utils';