+++ src/frontend/js/components/DashboardCourseList/index.spec.tsx
@@ -1,0 +1,89 @@
+import { MemoryRouter } from 'react-router-dom';
+import { QueryClientProvider } from '@tanstack/react-query';
+import { render, screen } from '@testing-library/react';
+import fetchMock from 'fetch-mock';
+import { IntlProvider } from 'react-intl';
+
+import {
+  RichieContextFactory as mockRichieContextFactory,
+  UserFactory,
+} from 'utils/test/factories/richie';
+import JoanieSessionProvider from 'contexts/SessionContext/JoanieSessionProvider';
+import { CourseListItemMock } from 'api/mocks/joanie/courses';
+import { CourseFactory } from 'utils/test/factories/joanie';
+import { TeacherCourseSearchFilters, CourseTypeFilter, CourseStatusFilter } from 'hooks/useCourses';
+import { createTestQueryClient } from 'utils/test/createTestQueryClient';
+import DashboardCourseList from '.';
+
+jest.mock('utils/context', () => ({
+  __esModule: true,
+  default: mockRichieContextFactory({
+    authentication: { backend: 'fonzie', endpoint: 'https://demo.endpoint' },
+    joanie_backend: { endpoint: 'https://joanie.endpoint' },
+  }).generate(),
+}));
+
+describe('components/DashboardCourseList', () => {
+  let nbApiCalls: number;
+  beforeEach(() => {
+    fetchMock.get('https://joanie.endpoint/api/v1.0/orders/', []);
+    fetchMock.get('https://joanie.endpoint/api/v1.0/credit-cards/', []);
+    fetchMock.get('https://joanie.endpoint/api/v1.0/addresses/', []);
+    nbApiCalls = 3;
+  });
+
+  it('do render', async () => {
+    CourseFactory.beforeGenerate((shape: CourseListItemMock) => {
+      return {
+        ...shape,
+        title: 'How to cook birds',
+      };
+    });
+    const courseCooking: CourseListItemMock = CourseFactory.generate();
+
+    CourseFactory.beforeGenerate((shape: CourseListItemMock) => {
+      return {
+        ...shape,
+        title: "Let's dance, the online leason",
+      };
+    });
+    const courseDancing: CourseListItemMock = CourseFactory.generate();
+    fetchMock.get('https://joanie.endpoint/api/v1.0/courses/?status=all&type=all', [
+      courseCooking,
+      courseDancing,
+    ]);
+
+    const filters: TeacherCourseSearchFilters = {
+      status: CourseStatusFilter.ALL,
+      type: CourseTypeFilter.ALL,
+    };
+
+    const user = UserFactory.generate();
+    render(
+      <IntlProvider locale="en">
+        <QueryClientProvider client={createTestQueryClient({ user })}>
+          <JoanieSessionProvider>
+            <MemoryRouter>
+              <DashboardCourseList
+                titleTranslated="DashboardCourseList test title"
+                filters={filters}
+              />
+            </MemoryRouter>
+          </JoanieSessionProvider>
+        </QueryClientProvider>
+      </IntlProvider>,
+    );
+    nbApiCalls += 1; // courses api call
+
+    expect(await screen.getByRole('heading', { name: /DashboardCourseList test title/ }));
+
+    const calledUrls = fetchMock.calls().map((call) => call[0]);
+    expect(calledUrls).toHaveLength(nbApiCalls);
+    expect(calledUrls).toContain('https://joanie.endpoint/api/v1.0/courses/?status=all&type=all');
+
+    expect(await screen.findByRole('heading', { name: /How to cook birds/ })).toBeInTheDocument();
+    expect(
+      screen.getByRole('heading', { name: /Let's dance, the online leason/ }),
+    ).toBeInTheDocument();
+  });
+});