+++ src/frontend/js/types/utils.ts
@@ -8,3 +8,9 @@ export type AddParameters<
   TFunction extends (...args: readonly unknown[]) => unknown,
   TParameters extends [...args: readonly unknown[]],
 > = (...args: [...Parameters<TFunction>, ...TParameters]) => ReturnType<TFunction>;
+
+export const isEnumValue = <T extends { [k: string]: string }>(
+  something: any,
+  enumObject: T,
+): something is T[keyof T] =>
+  typeof something === 'string' && Object.values(enumObject).includes(something);